+++ tests/test_cli.py
@@ -92,7 +92,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        \x1b[32m-2\x1b[39m      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
 dummy/dummy2    +2       \x1b[31m+1\x1b[39m      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
 dummy/dummy3    +2       \x1b[31m+2\x1b[39m      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
-TOTAL           -1       \x1b[32m-4\x1b[39m      +31.06%
+TOTAL           +4       \x1b[31m+1\x1b[39m      +15.00%
 """
     else:
         assert result.output == """\
@@ -102,7 +102,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        -2      +40.00%  -5, -6
 dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
 dummy/dummy3    +2       +2      -        +1, +2
-TOTAL           -1       -4      +31.06%
+TOTAL           +4       +1      +15.00%
 """
 
 
@@ -125,7 +125,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        \x1b[32m-2\x1b[39m      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
 dummy/dummy2    +2       \x1b[31m+1\x1b[39m      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
 dummy/dummy3    +2       \x1b[31m+2\x1b[39m      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
-TOTAL           -1       \x1b[32m-4\x1b[39m      +31.06%
+TOTAL           +4       \x1b[31m+1\x1b[39m      +15.00%
 """
         else:
             assert result.output == """\
@@ -135,7 +135,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        -2      +40.00%  -5, -6
 dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
 dummy/dummy3    +2       +2      -        +1, +2
-TOTAL           -1       -4      +31.06%
+TOTAL           +4       +1      +15.00%
 """
 
 
@@ -161,7 +161,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        -2      +40.00%  -5, -6
 dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
 dummy/dummy3    +2       +2      -        +1, +2
-TOTAL           -1       -4      +31.06%"""
+TOTAL           +4       +1      +15.00%"""
 
 
 # FIXME: when Click 4 is available, uncomment this.