+++ tests/test_reporters.py
@@ -76,7 +76,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        \x1b[32m-2\x1b[39m      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
 dummy/dummy2    +2       \x1b[31m+1\x1b[39m      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
 dummy/dummy3    +2       \x1b[31m+2\x1b[39m      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
-TOTAL           -1       \x1b[32m-4\x1b[39m      +31.06%"""
+TOTAL           +4       \x1b[31m+1\x1b[39m      +15.00%"""
 
 
 def test_text_report_delta__colorize_True__with_missing_range():
@@ -94,7 +94,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        \x1b[32m-2\x1b[39m      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
 dummy/dummy2    +2       \x1b[31m+1\x1b[39m      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
 dummy/dummy3    +2       \x1b[31m+2\x1b[39m      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
-TOTAL           -1       \x1b[32m-4\x1b[39m      +31.06%"""
+TOTAL           +4       \x1b[31m+1\x1b[39m      +15.00%"""
 
 
 def test_text_report_delta__colorize_False():
@@ -112,7 +112,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        -2      +40.00%  -5, -6
 dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
 dummy/dummy3    +2       +2      -        +1, +2
-TOTAL           -1       -4      +31.06%"""
+TOTAL           +4       +1      +15.00%"""
 
 
 def test_html_report():
@@ -243,7 +243,7 @@ dummy/__init__  -        -       -
 dummy/dummy     -        -2      +40.00%
 dummy/dummy2    +2       +1      -25.00%
 dummy/dummy3    +2       +2      -
-TOTAL           -1       -4      +31.06%"""
+TOTAL           +4       +1      +15.00%"""
 
 
 def test_html_report_delta__no_source():
@@ -303,9 +303,9 @@ def test_html_report_delta__no_source():
         <tfoot>
           <tr>
             <td>TOTAL</td>
-            <td>-1</td>
-            <td><span class="green">-4</span></td>
-            <td>+31.06%</td>
+            <td>+4</td>
+            <td><span class="red">+1</span></td>
+            <td>+15.00%</td>
           </tr>
         </tfoot>
       </table>
@@ -382,9 +382,9 @@ def test_html_report_delta():
         <tfoot>
           <tr>
             <td>TOTAL</td>
-            <td>-1</td>
-            <td><span class="green">-4</span></td>
-            <td>+31.06%</td>
+            <td>+4</td>
+            <td><span class="red">+1</span></td>
+            <td>+15.00%</td>
             <td></td>
           </tr>
         </tfoot>