+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* fix incorrect "TOTAL" row counts of the diff command when classes were added
+  or removed from the second report.
+
 ## 0.5.1 (2015-01-08)
 
 * Options of pycobertura diff `--missed` and `--no-missed` have been renamed to